selecting an item from the location or movement combobox now automatically adds it to the selected options list
also did some rearranging/modularizing of shared loc/mov code
