+++ src/main/python/models/shared_models.py
@@ -1,0 +1,41 @@
+
+from PyQt5.QtCore import (
+    Qt,
+    QSortFilterProxyModel
+)
+
+from constant import userdefinedroles as udr
+
+
+# used by both Location and Movement to filter selected tree paths (for combobox and selected paths list)
+class TreePathsProxyModel(QSortFilterProxyModel):
+
+    def __init__(self, parent=None, wantselected=None):
+        super(TreePathsProxyModel, self).__init__(parent)
+        self.wantselected = wantselected
+        self.setSortCaseSensitivity(Qt.CaseInsensitive)
+
+    def filterAcceptsRow(self, source_row, source_parent):
+        if self.wantselected is None:
+            source_index = self.sourceModel().index(source_row, 0, source_parent)
+            text = self.sourceModel().index(source_row, 0, source_parent).data()
+            return True
+        else:
+            source_index = self.sourceModel().index(source_row, 0, source_parent)
+            isselected = source_index.data(role=Qt.UserRole+udr.selectedrole)
+            path = source_index.data(role=Qt.UserRole+udr.pathdisplayrole)
+            text = self.sourceModel().index(source_row, 0, source_parent).data()
+            return self.wantselected == isselected
+
+    def updatesorttype(self, sortbytext=""):
+        if "alpha" in sortbytext and "path" in sortbytext:
+            self.setSortRole(Qt.DisplayRole)
+            self.sort(0)
+        elif "alpha" in sortbytext and "node" in sortbytext:
+            self.setSortRole(Qt.UserRole+udr.nodedisplayrole)
+            self.sort(0)
+        elif "tree" in sortbytext:
+            self.sort(-1)  # returns to sort order of underlying model
+        elif "select" in sortbytext:
+            self.setSortRole(Qt.UserRole+udr.timestamprole)
+            self.sort(0)
